requirements.txt
Django==4.2
djangorestframework==3.16.1
django-filter==25.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0